requirements.txt
altair==5.1.2
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.44.0
frozendict==2.3.8
gitdb==4.0.11
GitPython==3.1.40
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
multitasking==0.0.11
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
peewee==3.17.0
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.0
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.29.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
webencodings==0.5.1
yfinance==0.2.31
zipp==3.17.0
